+++ src/System/Taffybar/Context.hs
@@ -190,7 +190,8 @@ buildBarWindow context barConfig = do
       show $ strutConfig barConfig
 
   window <- Gtk.windowNew Gtk.WindowTypeToplevel
-  box <- Gtk.boxNew Gtk.OrientationHorizontal $ fromIntegral $ widgetSpacing barConfig
+  box <- Gtk.boxNew Gtk.OrientationHorizontal $ fromIntegral $
+         widgetSpacing barConfig
   _ <- widgetSetClassGI box "taffy-box"
   centerBox <- Gtk.boxNew Gtk.OrientationHorizontal $
                fromIntegral $ widgetSpacing barConfig